Enable JSON tests against Azure SqlDb
