+++ src/Microsoft.Data.SqlClient/tests/ManualTests/DataCommon/DataTestUtility.cs
@@ -92,9 +92,11 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
         //SQL Server EngineEdition
         private static string s_sqlServerEngineEdition;
 
+        // Currently, only Azure SQL supports vectors and JSON.
+        // Our CI images with specific SQL Server versions lag
+        // behind with vector and JSON support.
         // JSON Column type
-        public static readonly bool IsJsonSupported = false;
-
+        public static readonly bool IsJsonSupported = !IsNotAzureServer();
         // VECTOR column type
         public static readonly bool IsVectorSupported = !IsNotAzureServer();
